Implemented dictionary loader closing #1. Moved file and made tests.
Implemented the dictionary loader which has two methods, one to read
in all the files in a folder, another to read in all the words into
a list.

Also moved the dictionary files to a distinct folder so that only
those files are read and not the other liscenses and such.

Implemented some tests to teest the dictionary loader as well. Adding
in some test files as well to make it easy to test that the loader
functions work correctly

Also updated the pom file to make sure java 11 is default.